Collect BandCamp links
Now stores URLs of suggestions as the whole URL. BandCamp doesn't have
YouTube IDs. This has the added benefit of allowing users to link to a
specific time in a YouTube video.

Updates the youtube regexp to a more general regexp and adds the
regexp for BandCamp links.

Changes the Play link in templates to just the value of
videoid, which is now the whole URL.